proto fs-op caching

+++ core/server/master/src/main/java/alluxio/master/file/FileSystemMasterClientServiceHandler.java
@@ -31,6 +31,7 @@ import alluxio.grpc.CreateFilePResponse;
 import alluxio.grpc.DeletePRequest;
 import alluxio.grpc.DeletePResponse;
 import alluxio.grpc.FileSystemMasterClientServiceGrpc;
+import alluxio.grpc.FileSystemMasterCommonPOptions;
 import alluxio.grpc.FreePRequest;
 import alluxio.grpc.FreePResponse;
 import alluxio.grpc.GetFilePathPRequest;
@@ -152,6 +153,10 @@ public final class FileSystemMasterClientServiceHandler
   public void completeFile(CompleteFilePRequest request,
       StreamObserver<CompleteFilePResponse> responseObserver) {
     RpcUtils.call(LOG, () -> {
+      if (checkOperationComplete(request.getOptions().getCommonOptions())) {
+        LOG.info("YASS! Returning from completeFile.");
+        return CompleteFilePResponse.getDefaultInstance();
+      }
       AlluxioURI pathUri = getAlluxioURI(request.getPath());
       mFileSystemMaster.completeFile(pathUri,
           CompleteFileContext.create(request.getOptions().toBuilder()));
@@ -164,6 +169,10 @@ public final class FileSystemMasterClientServiceHandler
       StreamObserver<CreateDirectoryPResponse> responseObserver) {
     CreateDirectoryPOptions options = request.getOptions();
     RpcUtils.call(LOG, () -> {
+      if (checkOperationComplete(request.getOptions().getCommonOptions())) {
+        LOG.info("YASS! Returning from createDirectory.");
+        return CreateDirectoryPResponse.getDefaultInstance();
+      }
       AlluxioURI pathUri = getAlluxioURI(request.getPath());
       mFileSystemMaster.createDirectory(pathUri, CreateDirectoryContext.create(options.toBuilder())
           .withTracker(new GrpcCallTracker(responseObserver)));
@@ -175,6 +184,12 @@ public final class FileSystemMasterClientServiceHandler
   public void createFile(CreateFilePRequest request,
       StreamObserver<CreateFilePResponse> responseObserver) {
     RpcUtils.call(LOG, () -> {
+      if (checkOperationComplete(request.getOptions().getCommonOptions())) {
+        LOG.info("YASS! Returning from createFile.");
+        alluxio.wire.FileInfo fileInfo = mFileSystemMaster
+            .getFileInfo(getAlluxioURI(request.getPath()), GetStatusContext.defaults());
+        return CreateFilePResponse.newBuilder().setFileInfo(GrpcUtils.toProto(fileInfo)).build();
+      }
       AlluxioURI pathUri = getAlluxioURI(request.getPath());
       return CreateFilePResponse.newBuilder()
           .setFileInfo(GrpcUtils.toProto(mFileSystemMaster.createFile(pathUri,
@@ -303,6 +318,10 @@ public final class FileSystemMasterClientServiceHandler
   @Override
   public void remove(DeletePRequest request, StreamObserver<DeletePResponse> responseObserver) {
     RpcUtils.call(LOG, () -> {
+      if (checkOperationComplete(request.getOptions().getCommonOptions())) {
+        LOG.info("YASS! Returning from remove.");
+        return DeletePResponse.getDefaultInstance();
+      }
       AlluxioURI pathUri = getAlluxioURI(request.getPath());
       mFileSystemMaster.delete(pathUri, DeleteContext.create(request.getOptions().toBuilder())
           .withTracker(new GrpcCallTracker(responseObserver)));
@@ -313,6 +332,10 @@ public final class FileSystemMasterClientServiceHandler
   @Override
   public void rename(RenamePRequest request, StreamObserver<RenamePResponse> responseObserver) {
     RpcUtils.call(LOG, () -> {
+      if (checkOperationComplete(request.getOptions().getCommonOptions())) {
+        LOG.info("YASS! Returning from rename.");
+        return RenamePResponse.getDefaultInstance();
+      }
       AlluxioURI srcPathUri = getAlluxioURI(request.getPath());
       AlluxioURI dstPathUri = getAlluxioURI(request.getDstPath());
       mFileSystemMaster.rename(srcPathUri, dstPathUri,
@@ -420,6 +443,11 @@ public final class FileSystemMasterClientServiceHandler
       final List<String> holders = mFileSystemMaster.getStateLockSharedWaitersAndHolders();
       return GetStateLockHoldersPResponse.newBuilder().addAllThreads(holders).build();
     }, "getStateLockHolders", "request=%s", responseObserver, request);
+  }
+
+  private boolean checkOperationComplete(FileSystemMasterCommonPOptions commonOpts) {
+    return commonOpts != null && commonOpts.hasOpId()
+        && mFileSystemMaster.isOpComplete(commonOpts.getOpId());
   }
 
   /**
